requirements.txt
# tensorflow==2.16.1
tflite-runtime==2.16.1
Pillow==10.0.0
numpy>=1.24.3,<2.0.0
matplotlib>=3.7.1
flask==3.0.2
requests==2.31.0